[Rust] Integer Complexity 1 - Prime numbers (Reddit)

+++ Reddit/Easy/integer_complexity_1/src/main.rs
@@ -7,8 +7,8 @@ use integer_complexity_1 as complexity;
 fn main() {
     let stdin = io::stdin();
 
-    let input: i32 = stdin.lock().lines()
-        .nth(0).unwrap().unwrap().parse::<i32>().unwrap();
+    let input: u64 = stdin.lock().lines()
+        .nth(0).unwrap().unwrap().parse::<u64>().unwrap();
 
     println!("{}", complexity::smallest_factor_sum(input));
 }
